otrrentetl/requirements.txt
aniso8601==1.3.0
appdirs==1.4.3
click==6.7
MarkupSafe==1.0
packaging==16.8
passlib==1.7.1
pip==9.0.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
setuptools==36.4.0
six==1.10.0
asn1crypto==0.22.0
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage==0.36.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
idna==2.6
pycparser==2.18
requests==2.18.4
urllib3==1.22
